+++ src/components/MainListTools/MainListTools.js
@@ -45,10 +45,7 @@ export default class MainListTools extends React.Component {
 
     fireAction(status) {
         const checked = this.props.checkedItems;
-        if (status === 'completed') {
-            alert('Are you sure you want to mark the selected items')
-        }
-        if (status === 'completed' || status === 'reached') {
+        if (status === 'reached') {
             checked.forEach(itemId => {
                 const item = this.context.listItems.find(item => item.id === itemId)
                 const foundPm = this.context.pms.find(pm => pm.pm_email === item.pm_email)
@@ -69,20 +66,50 @@ export default class MainListTools extends React.Component {
                     .then(this.props.clearChecked())
                     .catch(error => console.log(error))
             })
+        } else if (status === 'completed') {
+            if (window.confirm('Are you sure you want to mark the checked items completed?')) {
+                checked.forEach(itemId => {
+                    const item = this.context.listItems.find(item => item.id === itemId)
+                    const foundPm = this.context.pms.find(pm => pm.pm_email === item.pm_email)
+                    const pmId = foundPm.id;
+                    const { project, advisor } = item;
+                    fetch(`${config.API_ENDPOINT}/list/${itemId}`, {
+                        method: 'PATCH',
+                        headers: {
+                            'content-type': 'application/json',
+                            'Authorization': `Bearer ${window.sessionStorage.getItem(config.TOKEN_KEY)}`
+                        },
+                        body: JSON.stringify({ status, project, advisor, pm_id: pmId })
+                    })
+                        .then(res => 
+                            (!res.ok)
+                                ? res.json().then(e => Promise.reject(e))
+                                : this.context.updateItemStatus(itemId, status))
+                        .then(this.props.clearChecked())
+                        .catch(error => console.log(error))
+                })
+            } else {
+                return null;
+            }  
         } else if (status === 'delete') {
-            checked.forEach(item => {
-                fetch(`${config.API_ENDPOINT}/list/${item}`, {
-                    method: 'DELETE',
-                    headers: {
-                        'content-type': 'application/json',
-                        'Authorization': `Bearer ${window.sessionStorage.getItem(config.TOKEN_KEY)}`
-                    }
+            if (window.confirm('Are you sure you want to delete the selected items?')) {
+                checked.forEach(item => {
+                    fetch(`${config.API_ENDPOINT}/list/${item}`, {
+                        method: 'DELETE',
+                        headers: {
+                            'content-type': 'application/json',
+                            'Authorization': `Bearer ${window.sessionStorage.getItem(config.TOKEN_KEY)}`
+                        }
+                            
+                    })
+                        .then(res => this.context.deleteItem(item))
+                        .then(this.props.clearChecked())
                         
                 })
-                    .then(res => this.context.deleteItem(item))
-                    .then(this.props.clearChecked())
-                    
-            })
+            } else {
+                return null;
+            }
+            
         } else if (status === 'reset') {
             this.context.listItems.forEach(item => {
                 const { project, advisor } = item
